feat: Add Alert helpers, export EventFilters, fix ndr-storage
Enhanced shared crates for service compatibility:

ndr-core improvements:
- Added Alert helper methods: source_ip(), dest_ip(), rule_name()
- Exported EventFilters from ports module

ndr-storage fixe:
- Fixed redis cache ttl type (usize → u64)
- Fixed OpenSearchEventStore to use EventFilters properly

Service migration progress:
- rust-alert-correlator engine.rs adapted to shared types

These changes enable backwards compatibility for services migrating
to shared Alert type while maintaining type safety.

+++ services/crates/ndr-storage/src/redis/cache.rs
@@ -26,7 +26,7 @@ impl RedisCache {
         &self,
         key: &str,
         value: &T,
-        ttl_seconds: usize,
+        ttl_seconds: u64,
     ) -> Result<()> {
         let mut conn = self.get_connection()?;
         let serialized = serde_json::to_string(value)
